save →　saveAndFlush

+++ RingishoSystem/src/main/java/com/kadai/domain/entity/App.java
@@ -80,7 +80,7 @@ public class App {
 		this.appId = appId;
 	}
 
-	public long getEmployeeId() {
+	public int getEmployeeId() {
 		return employeeId;
 	}
